Make input configuration options available
